+++ pmd-core/src/test/resources/net/sourceforge/pmd/cpd/files/file_with_utf8_bom.java
@@ -1,1 +1,8 @@
-﻿
+﻿/**
+ * This file is using UTF-8 with BOM encoding.
+ *
+ * ä
+ */
+public class FileWith_UTF-8-BOM_Encoding {
+
+}